refactor(soffit-pwa): manager folder structure
