--- uniquely/soffit-pwa/src/director/index.ts
@@ -1,4 +1,0 @@
-import './change-language.js';
-import './content-provider.js';
-import './l18e-loader.js';
-import './submit-form-command-handler.js';
