--- uniquely/soffit-pwa/src/director/logger.ts
@@ -1,3 +1,0 @@
-import {createLogger} from '@alwatr/logger';
-
-export const logger = createLogger('soffit-director');
