+++ uniquely/soffit-pwa/src/manager/index.ts
@@ -1,0 +1,5 @@
+import './change-language.js';
+import './context-provider/home-page-content.js';
+import './context-provider/product-page-content.js';
+import './l18e.js';
+import './submit-form-command-handler.js';
