+++ uniquely/soffit-pwa/src/manager/logger.ts
@@ -1,0 +1,3 @@
+import {createLogger} from '@alwatr/logger';
+
+export const logger = createLogger('soffit-manager');
